Fix bugs in class GfCheckboxValidators
It was found that using ajax to load the form causes validation to break down because the class bases initialization only on page loading. Accordingly, rewrite the class to also depend on event bubbling, including methods GfCheckboxValidators, finishHidingValidators, initValidation, and IsObjValid, so validation will properly function when the form is loaded and updated via ajax.