Remove application of removed GfCheckboxValidators module
Removed module was causing accessibility problems under the following specific and unforseen condition. When save and continue links are being used on a graviy form and the user saves the form to continue it later from the page the validated checkboxes are on, the checkbox validator field does not properly keep its validated value after the user accesses the link again in the future. Moreover, the error messaging provided by the valdiation scheme proved not to be helpful enough in practice.